requirements.txt
Django==4.1.7
asgiref ==3.6.0
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.3
wheel==0.38.4
djangorestframework==3.14.0